+++ events/src/lib.rs
@@ -21,7 +21,7 @@ pub mod proto {
 }
 
 #[cfg(feature = "grpc")]
-pub use grpc::{GrpcEventSink, GrpcEventSinkError, GrpcLimits};
+pub use grpc::{GrpcEventSink, GrpcEventSinkError, GrpcLimits, GrpcTlsParams};
 
 #[cfg(feature = "zmq")]
 pub mod zmq;